+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes to `drupal/opening-hours` module.
 
+## Unreleased
+
+### Added
+
+* VG-1385: Added keyboard support for the month view.
+
 ## [8.x-1.0-alpha12]
 
 ### Fixed